@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Socket } from 'ng-socket-io';
-import 'rxjs/add/operator/map';
+
 
 @Injectable()
 export class AvanzadoService {
